requirements.txt
fastapi
uvicorn
python-multipart
streamlit
transformers
torch
Pillow
requests